add fixes, remove duplicate imports
