+++ portfolio/src/main/webapp/chartPage.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <title>My Portfolio</title>
     <script src="https://www.gstatic.com/charts/loader.js"></script>
-    <script src="script.js"></script>
+    <script src="chartScript.js"></script>
     <link rel="stylesheet" href="style.css">
   </head>
   <body>
